refactor(ui): split chatSessions.ts god object into modules
