+++ scripts/.god-object-baseline.txt
@@ -4,6 +4,5 @@
 1106 ui/src/routes/instances.$id.tsx
 1065 sandbox-runtime/src/session_auth.rs
 738 ui/src/routes/sandboxes.$id.tsx
-701 ui/src/lib/stores/chatSessions.ts
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
 606 sandbox-runtime/src/util.rs
